fix: Add marker labels and circles in to map in opposite orders.
I want the first labelled point to be the topmost marker on the map. Map markers are drawn in the
order they're added, consequently the first marker added is the lowest one and the last marker the
topmost. This feels kind of counterintuitive, so I've switched things around to add the labels, then
added the markers in the reverse order.